pyproject.toml
[tool.poetry]
name = "errbot-slack-bolt-backend"
version = "0.1.0"
description = "Sladk backend by Slack Bolt"
authors = ["Kazuya Takei <myself@attakei.net>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.9"
slack-bolt = ">1.0.0"

[tool.poetry.dev-dependencies]
errbot = "^6.1.7"
python-dotenv = "^0.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
